+++ starsky/starsky/clientapp/src/pages/import-page.tsx
@@ -7,14 +7,14 @@ import MenuSearch from '../components/menu-search';
 
 const ImportPage: FunctionComponent<RouteComponentProps> = (props) => {
 
-  return (<DropArea><div>
+  return (<div>
     <MenuSearch></MenuSearch>
     <div className="content">
       <div className="content--header">Importeren</div>
       <div className="content--subheader"><a href="/v1/import"><u>Test functionaliteit: Ga naar oude weergave</u></a></div>
+      <DropArea enableInputField={true} enableDragAndDrop={true}></DropArea>
     </div>
   </div>
-  </DropArea>
   )
 }